changed file names and process files in order

+++ code/sidecars/generate_sidecars.py
@@ -174,40 +174,40 @@ def main(args):
                         if bk2_file != "Missing file" and type(bk2_file) != float:
                             print("Adding : " + bk2_file)
                             if op.exists(bk2_file):
+                                if not op.exists(bk2_file.replace(".bk2", ".json")):
+                                    # replay and save using retro legacy function
+                                    # This bloc is important to get the skip_first_step right before playback_movie
+                                    game = None
+                                    scenario = None
+                                    inttype = retro.data.Integrations.CUSTOM_ONLY
+                                    movie = retro.Movie(bk2_file)
+                                    skip_first_step = bk2_idx==0
+                                    if game == None:
+                                        game = movie.get_game()
+                                    emulator = retro.make(game, scenario=scenario, inttype=inttype, render_mode=False)
+                                    emulator.initial_state = movie.get_state()
+                                    emulator.reset()
+                                    if skip_first_step:
+                                        movie.step()
 
-                                # replay and save using retro legacy function
-                                # This bloc is important to get the skip_first_step right before playback_movie
-                                game = None
-                                scenario = None
-                                inttype = retro.data.Integrations.CUSTOM_ONLY
-                                movie = retro.Movie(bk2_file)
-                                skip_first_step = bk2_idx==0
-                                if game == None:
-                                    game = movie.get_game()
-                                emulator = retro.make(game, scenario=scenario, inttype=inttype, render_mode=False)
-                                emulator.initial_state = movie.get_state()
-                                emulator.reset()
-                                if skip_first_step:
-                                    movie.step()
+                                    npy_file = bk2_file.replace(".bk2", ".npz")
+                                    video_file = bk2_file.replace(".bk2", ".mp4")
+                                    playback_movie(emulator, movie, npy_file=npy_file, video_file=video_file, lossless='mp4', info_file=True)
+                                    emulator.close()
 
-                                npy_file = bk2_file.replace(".bk2", ".npz")
-                                video_file = bk2_file.replace(".bk2", ".mp4")
-                                playback_movie(emulator, movie, npy_file=npy_file, video_file=video_file, lossless='mp4', info_file=True)
-                                emulator.close()
+                                    repvars = format_repvars_dict(bk2_file, emulator)
 
-                                repvars = format_repvars_dict(bk2_file, emulator)
+                                    info_dict = create_info_dict(repvars)
 
-                                info_dict = create_info_dict(repvars)
+                                    # write info_dict to json file
+                                    json_sidecar_fname = bk2_file.replace(".bk2", ".json")
+                                    with open(json_sidecar_fname, 'w') as f:
+                                        json.dump(info_dict, f)
 
-                                # write info_dict to json file
-                                json_sidecar_fname = bk2_file.replace(".bk2", ".json")
-                                with open(json_sidecar_fname, 'w') as f:
-                                    json.dump(info_dict, f)
-
-                                # write repvars dict as pkl
-                                pkl_sidecar_fname = bk2_file.replace(".bk2", ".pkl")
-                                with open(pkl_sidecar_fname, 'wb') as f:
-                                    pickle.dump(repvars, f)
+                                    # write repvars dict as pkl
+                                    pkl_sidecar_fname = bk2_file.replace(".bk2", ".pkl")
+                                    with open(pkl_sidecar_fname, 'wb') as f:
+                                        pickle.dump(repvars, f)
                                     
 if __name__ == "__main__":
     args = parser.parse_args()
